@@ -1,69 +1,192 @@
 {
-  "name": "smart_monitor",
-  "version": "1.0.0",
-  "description": "Smart Monitor: A Digital Signage MQTT IoT Based Solution Server",
-  "main": "./dist/server.js",
-  "scripts": {
-    "build": "tsc",
-    "dev": "ts-node ./lib/server.ts",
-    "start": "nodemon -e js ./dist/server.js",
-    "deploy": "node ./dist/server.js",
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "author": "Pratik Chowdhury (pratikc@live.co.uk)",
-  "license": "BSD-2-Clause",
-  "dependencies": {
-    "@types/ejs": "^3.0.4",
-    "@types/express-session": "^1.17.0",
-    "@types/fluent-ffmpeg": "^2.1.14",
-    "@types/mime": "^2.0.3",
-    "@types/mongodb": "^3.5.25",
-    "@types/multer": "^1.4.3",
-    "@types/passport": "^1.0.4",
-    "@types/passport-local": "^1.0.33",
-    "@types/pg": "^7.14.4",
-    "@types/session-file-store": "^1.2.1",
-    "@types/sharp": "^0.25.1",
-    "@types/validator": "^13.1.0",
-    "bcrypt": "^5.0.0",
-    "body-parser": "^1.19.0",
-    "connect-session-sequelize": "^7.0.1",
-    "cookie-parser": "^1.4.5",
-    "cors": "^2.8.5",
-    "dotenv": "^8.2.0",
-    "ejs": "^3.1.3",
-    "express": "^4.17.1",
-    "express-session": "^1.17.1",
-    "express-validator": "^6.6.0",
-    "fluent-ffmpeg": "^2.1.2",
-    "gif-info": "^1.0.1",
-    "jsonwebtoken": "^8.5.1",
-    "mime": "^2.4.6",
-    "mongodb": "^3.5.9",
-    "mqtt": "^4.1.0",
-    "multer": "^1.4.2",
-    "passport": "^0.4.1",
-    "passport-local": "^1.0.0",
-    "pg": "^8.3.0",
-    "reflect-metadata": "^0.1.13",
-    "sequelize": "^6.1.1",
-    "sequelize-typescript": "next",
-    "session-file-store": "^1.4.0",
-    "sharp": "^0.25.4"
-  },
-  "devDependencies": {
-    "@types/bcrypt": "^3.0.0",
-    "@types/body-parser": "^1.19.0",
-    "@types/cookie-parser": "^1.4.2",
-    "@types/cors": "^2.8.6",
-    "@types/express": "^4.17.7",
-    "@types/jsonwebtoken": "^8.5.0",
-    "@types/node": "^14.0.26",
-    "nodemon": "^2.0.4",
-    "npm-check-updates": "^7.0.2",
-    "typescript": "^3.9.7"
-  },
-  "engines": {
-    "node": ">=10.0.0"
-  }
+   "name": "smart_monitor",
+   "version": "1.0.0",
+   "description": "Smart Monitor: A Digital Signage MQTT IoT Based Solution Server",
+   "main": "./dist/server.js",
+   "scripts": {
+      "build": "tsc",
+      "dev": "ts-node ./lib/server.ts",
+      "start": "nodemon -e js ./dist/server.js",
+      "deploy": "node ./dist/server.js",
+      "lint": "eslint --cache --max-warnings=0 --fix lib",
+      "format": "prettier --write lib",
+      "test": "echo \"Error: no test specified\" && exit 1"
+   },
+   "author": "Pratik Chowdhury (pratikc@live.co.uk)",
+   "license": "BSD-2-Clause",
+   "dependencies": {
+      "@types/ejs": "^3.0.4",
+      "@types/express-session": "^1.17.0",
+      "@types/fluent-ffmpeg": "^2.1.14",
+      "@types/mime": "^2.0.3",
+      "@types/mongodb": "^3.5.26",
+      "@types/multer": "^1.4.4",
+      "@types/passport": "^1.0.4",
+      "@types/passport-local": "^1.0.33",
+      "@types/pg": "^7.14.4",
+      "@types/session-file-store": "^1.2.1",
+      "@types/sharp": "^0.25.1",
+      "@types/validator": "^13.1.0",
+      "bcrypt": "^5.0.0",
+      "body-parser": "^1.19.0",
+      "connect-session-sequelize": "^7.0.1",
+      "cookie-parser": "^1.4.5",
+      "cors": "^2.8.5",
+      "dotenv": "^8.2.0",
+      "ejs": "^3.1.5",
+      "express": "^4.17.1",
+      "express-session": "^1.17.1",
+      "express-validator": "^6.6.1",
+      "fluent-ffmpeg": "^2.1.2",
+      "gif-info": "^1.0.1",
+      "jsonwebtoken": "^8.5.1",
+      "mime": "^2.4.6",
+      "mongodb": "^3.6.0",
+      "mqtt": "^4.2.1",
+      "multer": "^1.4.2",
+      "passport": "^0.4.1",
+      "passport-local": "^1.0.0",
+      "pg": "^8.3.3",
+      "reflect-metadata": "^0.1.13",
+      "sequelize": "^6.1.1",
+      "sequelize-typescript": "next",
+      "session-file-store": "^1.4.0",
+      "sharp": "^0.26.0"
+   },
+   "devDependencies": {
+      "@types/bcrypt": "^3.0.0",
+      "@types/body-parser": "^1.19.0",
+      "@types/cookie-parser": "^1.4.2",
+      "@types/cors": "^2.8.7",
+      "@types/express": "^4.17.7",
+      "@types/jsonwebtoken": "^8.5.0",
+      "@types/node": "^14.6.1",
+      "@typescript-eslint/eslint-plugin": "^3.10.1",
+      "@typescript-eslint/parser": "^3.10.1",
+      "eslint": "^7.7.0",
+      "eslint-config-airbnb-typescript": "^9.0.0",
+      "eslint-config-prettier": "^6.11.0",
+      "eslint-plugin-eslint-comments": "^3.2.0",
+      "eslint-plugin-import": "^2.22.0",
+      "eslint-plugin-promise": "^4.2.1",
+      "eslint-plugin-svelte3": "^2.7.3",
+      "husky": "^4.2.5",
+      "lint-staged": "^10.2.13",
+      "nodemon": "^2.0.4",
+      "prettier": "^2.1.1",
+      "svelte": "^3.24.1",
+      "typescript": "^4.0.2"
+   },
+   "prettier": {
+      "trailingComma": "none",
+      "printWidth": 120,
+      "proseWrap": "preserve",
+      "semi": true,
+      "singleQuote": true,
+      "useTabs": false,
+      "tabWidth": 3,
+      "arrowParens": "avoid"
+   },
+   "eslintConfig": {
+      "root": true,
+      "parser": "@typescript-eslint/parser",
+      "parserOptions": {
+         "tsconfigRootDir": ".",
+         "project": [
+            "tsconfig.json"
+         ]
+      },
+      "rules": {
+         "@typescript-eslint/explicit-module-boundary-types": "off",
+         "object-shorthand": [
+            "error",
+            "never"
+         ],
+         "import/no-extraneous-dependencies": [
+            "error",
+            {
+               "devDependencies": true
+            }
+         ],
+         "no-return-await": "off",
+         "import/newline-after-import": "off",
+         "import/no-mutable-exports": 0,
+         "no-labels": 0,
+         "no-prototype-builtins": "off",
+         "no-console": "off",
+         "no-restricted-syntax": 0,
+         "max-classes-per-file": [
+            "error",
+            4
+         ],
+         "@typescript-eslint/no-namespace": "off",
+         "@typescript-eslint/lines-between-class-members": [
+            "error",
+            "always",
+            {
+               "exceptAfterSingleLine": true
+            }
+         ],
+         "lines-between-class-members": [
+            "error",
+            "always",
+            {
+               "exceptAfterSingleLine": true
+            }
+         ]
+      },
+      "plugins": [
+         "@typescript-eslint",
+         "svelte3"
+      ],
+      "extends": [
+         "airbnb-typescript/base",
+         "plugin:@typescript-eslint/recommended",
+         "plugin:eslint-comments/recommended",
+         "plugin:promise/recommended",
+         "prettier",
+         "prettier/@typescript-eslint"
+      ],
+      "overrides": [
+         {
+            "files": [
+               "**/*.svelte"
+            ],
+            "processor": "svelte3/svelte3"
+         }
+      ]
+   },
+   "engines": {
+      "node": ">=10.0.0"
+   },
+   "husky": {
+      "hooks": {
+         "pre-commit": "lint-staged",
+         "pre-push": "lint-staged"
+      }
+   },
+   "lint-staged": {
+      "*.ts": [
+         "npm run lint"
+      ],
+      "*.{md,json,yml,ts}": [
+         "npm run format"
+      ]
+   },
+   "settings": {
+      "import/resolver": {
+         "node": {
+            "extensions": [
+               ".MD",
+               ".ts",
+               ".json"
+            ]
+         }
+      },
+      "import/extensions": [
+         ".MD",
+         ".ts",
+         ".json"
+      ]
+   }
 }
