@@ -1,27 +1,7 @@
-import { networkInterfaces } from "os";
-
 export function GetIP() {
-  const interfaces = networkInterfaces();
-  for (const devName in interfaces) {
-    const iface = interfaces[devName];
-    if (iface != null)
-      for (let i = 0; i < iface.length; i++) {
-        const alias = iface[i];
-        if (
-          alias.family === "IPv4" &&
-          alias.address !== "127.0.0.1" &&
-          !alias.internal
-        )
-          return alias.address;
-      }
-  }
-
-  return "localhost";
+   return '0.0.0.0';
 }
 
-export function IfDockerisedOrSelectDefault(
-  dockerValue: any,
-  defValue: string
-) {
-  return process.env.APP_IS_DOCKERISED ? String(dockerValue) : defValue;
+export function IfDockerisedOrSelectDefault(dockerValue: string, defValue: string) {
+   return process.env.APP_IS_DOCKERISED ? dockerValue : defValue;
 }
