@@ -1,3 +1,3 @@
-export { Users } from "./Users.Routes";
-export { Displays } from "./Display.Routes";
-export { Files } from "./Files.Routes";
+export { Users } from './Users.Routes';
+export { Displays } from './Display.Routes';
+export { Files } from './Files.Routes';
